# specs/001-buildcare-ai-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Buildcare AI API
  version: 0.1.0
  description: |
    Buildcare AI 백엔드 REST 계약. 근거: specs/001-buildcare-ai-platform/spec.md (FR-xxx),
    Design/SD_01 (P0~P8, G0~G10), Design/SD_02 (S1~S8B), Design/buildcare_ddl.sql.

    공통 규칙
    - 인증: httpOnly 쿠키 `bc_session`(JWT). 비회원은 `bc_device` 쿠키. 상태 변경 요청은 `X-CSRF-Token` 필요.
    - 게이트 미통과는 항상 `GateBlock` 본문으로 응답한다(FR-004). 문구는 gate_def.block_message 와 같다.
      HTTP 상태: G0 미로그인 401 · G0 권한 없음 403 · G1 402 · G2 402 · G3 422 · G5 422 · G4 502 · G6 409 · G7 200(blocked 필드) · G8 409 · G9 409 · G10 409
    - 결제·분석 요청·기록 저장·요청 확정은 `Idempotency-Key` 헤더를 받는다(중복 제출 방지).
    - 개인정보는 마스킹 필드만 반환한다(FR-090).
servers:
  - url: http://localhost:3000
    description: 로컬 개발 백엔드
security:
  - sessionCookie: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      security: []
      summary: 이메일·비밀번호 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: 로그인 성공. 세션 쿠키 발급. 이 기기에 연결 가능한 비회원 분석이 있으면 guestLinkable=true
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Me'
                  - type: object
                    properties:
                      guestLinkable: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/auth/me:
    get:
      tags: [auth]
      security: [{ sessionCookie: [] }, { deviceCookie: [] }]
      summary: 현재 사용자(비회원 포함)와 역할·메뉴
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
          description: OK
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/buildings:
    get:
      tags: [building]
      summary: 권한 범위 건물 목록 C8 (v_access_check)
      parameters:
        - name: access
          in: query
          schema: { type: string, enum: [record, manage] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BuildingSummary' }
  /api/buildings/{buildingId}/rail:
    get:
      tags: [building]
      summary: 건물 레일 (이력·지연·연결 대기 건수 — v_building_rail)
      parameters: [{ $ref: '#/components/parameters/BuildingId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BuildingRail' }
        '403': { $ref: '#/components/responses/Gate' }
  /api/buildings/{buildingId}/history:
    get:
      tags: [building]
      summary: 시간순 통합 이력 (v_building_history) — 조회 전용 (FR-055)
      parameters:
        - $ref: '#/components/parameters/BuildingId'
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: defectType, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK. 연동 실패 시 sourceStatus.externalIncluded=false (FR-052)
          content:
            application/json:
              schema:
                type: object
                properties:
                  sourceStatus: { $ref: '#/components/schemas/SourceStatus' }
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/HistoryEntry' }
        '403': { $ref: '#/components/responses/Gate' }
  /api/buildings/{buildingId}/repeat-defects:
    get:
      tags: [building]
      summary: 반복 하자 패턴 (v_repeat_defect, G7 = v_pattern_gate)
      parameters: [{ $ref: '#/components/parameters/BuildingId' }]
      responses:
        '200':
          description: G7 미통과면 blocked 에 GateBlock, patterns 는 빈 배열
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocked:
                    oneOf:
                      - $ref: '#/components/schemas/GateBlock'
                      - type: 'null'
                  patterns:
                    type: array
                    items: { $ref: '#/components/schemas/RepeatPattern' }
  /api/buildings/{buildingId}/bms-sync:
    post:
      tags: [building]
      summary: 건물관리시스템 동기화 수동 실행 (bms_sync_run)
      parameters: [{ $ref: '#/components/parameters/BuildingId' }]
      responses:
        '200':
          description: 결과(ok/failed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceStatus' }
  /api/permission-requests:
    post:
      tags: [building]
      summary: G0 [권한 요청 보내기] (FR-003)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestedScreen]
              properties:
                buildingId: { type: integer }
                orgId: { type: integer }
                requestedScreen: { type: string, enum: [S3, S4, S5, S6, S7A, S7B, S8A, S8B] }
      responses:
        '201': { description: 요청 기록 }

  # ---------------------------------------------------------------- record (S3)
...
  /api/admin/constants/{key}:
    put:
      tags: [admin]
      summary: 운영 기준값 설정 (null = 보수적 동작)
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: string, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Constant' }
  /api/admin/notices:
    post:
      tags: [admin]
      summary: 고지 문구 새 판본 (기존 결과는 당시 판본 유지 — FR-018)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, body]
              properties:
                kind: { type: string, enum: [analysis, priority, share] }
                body: { type: string, maxLength: 1000 }
      responses:
        '201':
          description: 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notice' }
  /api/admin/permission-requests/{permReqId}:
    post:
      tags: [admin]
      summary: 권한 요청 처리 (기업 관리자 — 자기 조직 건물 / 운영자)
      parameters:
        - name: permReqId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution: { type: string, enum: [granted, rejected] }
                accessKind: { type: string, enum: [record, manage] }
      responses:
        '200': { description: 처리 완료 }

  # ---------------------------------------------------------------- 구현 단계 추가 (2026-10-03, research R22 RBAC 포함)
...
  /api/auth/password:
    post:
      tags: [auth]
      summary: 비밀번호 변경 (token_version+1 — 다른 세션 즉시 무효)
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [currentPassword, newPassword], properties: { currentPassword: { type: string }, newPassword: { type: string, minLength: 10 } } } } }
      responses:
        '200': { description: 변경됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: 비밀번호 정책 위반 }
...
  /api/buildings/{buildingId}:
    get:
      tags: [building]
      summary: 건물 정보(유형 읽기 전용 표시용)
      parameters: [{ name: buildingId, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Gate' } }
  /api/buildings/{buildingId}/history/{kind}/{sourceId}:
    get:
      tags: [building]
      summary: 이력 항목 상세(사진·AI 결과+고지·판정·보수 결과) — 읽기 전용
      parameters:
        - { name: buildingId, in: path, required: true, schema: { type: integer } }
        - { name: kind, in: path, required: true, schema: { type: string, enum: [ai_analysis, field_record, expert_verdict, expert_connection, external] } }
        - { name: sourceId, in: path, required: true, schema: { type: integer } }
      responses: { '200': { description: OK } }
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: 계정 목록(마스킹) — 운영자 전체 · 기업 관리자 소속 조직
      parameters:
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled] } }
        - { name: q, in: query, schema: { type: string } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AdminUser' } } } } } }
    post:
      tags: [admin]
      summary: 계정 생성 — 임시 비밀번호 1회 반환, 첫 로그인 시 변경 강제
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [email, displayName, roles], properties: { email: { type: string, format: email }, displayName: { type: string }, phone: { type: string, nullable: true }, orgId: { type: integer, nullable: true }, roles: { type: array, items: { type: string } } } } } }
      responses: { '201': { description: 생성 } }
  /api/admin/users/{userId}:
    get:
      tags: [admin]
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: OK } }
    patch:
      tags: [admin]
      summary: 비활성화/재활성화 (token_version+1)
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { disabled: { type: boolean } } } } } }
      responses: { '200': { description: OK } }
  /api/admin/users/{userId}/roles:
    put:
      tags: [admin]
      summary: 역할 설정(업무 5종 + operator) — 마지막 운영자 보호, 세션 즉시 무효화
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RoleAssignment' } } } }
      responses: { '200': { description: OK }, '409': { description: 마지막 운영자 } }
  /api/admin/users/{userId}/reset-password:
    post:
      tags: [admin]
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: 임시 비밀번호 1회 반환 } }
  /api/admin/users/{userId}/building-access:
    post:
      tags: [admin]
      summary: 건물 권한 부여 — 운영자 전체, 기업 관리자 소속 조직 건물만
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BuildingAccessGrant' } } } }
      responses: { '201': { description: OK }, '403': { description: 조직 밖 건물 } }
    delete:
      tags: [admin]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer } }
        - { name: buildingId, in: query, required: true, schema: { type: integer } }
        - { name: accessKind, in: query, required: true, schema: { type: string, enum: [record, manage] } }
      responses: { '200': { description: OK } }
  /api/admin/permission-requests:
    get:
      tags: [admin]
      parameters: [{ name: status, in: query, schema: { type: string, enum: [pending, all] } }]
      responses: { '200': { description: OK } }
...
  /api/admin/rbac/roles/{role}/permissions:
    put:
      tags: [admin]
      summary: 역할 권한 변경(운영자) — 잠금 권한은 운영자에게서 제거 불가
      parameters: [{ name: role, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { permissions: { type: array, items: { type: string } } } } } } }
      responses: { '200': { description: OK }, '409': { description: 잠금 권한 } }
...
  /api/admin/plans/{planCode}/price:
    put:
      tags: [admin]
      summary: 요금제 금액 설정(NULL = 미설정, 결제 불가)
      parameters: [{ name: planCode, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bc_session
    deviceCookie:
      type: apiKey
      in: cookie
      name: bc_device
  parameters:
...
    BuildingId: { name: buildingId, in: path, required: true, schema: { type: integer } }
...
  responses:
    Gate:
      description: 게이트 미통과 — C2 차단 블록으로 표시
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GateBlock' }
    Unauthorized:
      description: G0 미로그인 — 로그인 후 같은 화면 복귀
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GateBlock' }
  schemas:
...
    GateBlock:
      type: object
      description: FR-004~FR-007. message 는 gate_def.block_message 와 동일
      required: [gate, message, releaseParty, actions]
      properties:
        gate: { type: string, enum: [G0, G1, G2, G3, G4, G5, G6, G7, G8, G9, G10] }
        message: { type: string }
        reason: { type: string, description: 구체 사유 (예 '사진이 흐립니다') }
        releaseParty: { type: string }
        selfRelease: { type: boolean, description: false 면 «요청 보내기» 행동만 (FR-006) }
        actions:
          type: array
          items:
            type: object
            properties:
              id: { type: string, example: retry-upload }
              label: { type: string, example: 사진 다시 올리기 }
              href: { type: string }
        missingFields:
          type: array
          items: { type: string, enum: [location, defectType, repairStatus, aiMatch, photo] }
        gateEventId: { type: integer }
...
    Code:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
    Me:
      type: object
      properties:
        userId: { type: integer }
        isGuest: { type: boolean }
        nameMasked: { type: string }
        emailMasked: { type: string }
        roles:
          type: array
          items: { type: string, enum: [guest, general, facility, building, enterprise, expert, operator] }
        permissions:
          type: array
          description: research R22 — 화면·버튼 판단에 쓰는 권한 코드(backend/src/auth/permissions.ts)
          items: { type: string }
        mustChangePassword: { type: boolean }
        orgId: { type: integer, nullable: true }
        menu:
          type: array
          items: { type: string, enum: [S1, S2, S3, S4, S5, S6, S7A, S7B, S8A, S8B, ADMIN] }
...
    Notice:
      type: object
      properties:
        noticeId: { type: integer }
        kind: { type: string, enum: [analysis, priority, share] }
        body: { type: string }
        effectiveFrom: { type: string, format: date-time }
...
    BuildingSummary:
      type: object
      properties:
        buildingId: { type: integer }
        buildingName: { type: string }
        buildingType: { $ref: '#/components/schemas/Code' }
        accessKinds: { type: array, items: { type: string, enum: [record, manage] } }
    BuildingRail:
      type: object
      properties:
        buildingId: { type: integer }
        buildingName: { type: string }
        historyCount: { type: integer }
        overdueCount: { type: integer }
        awaitingCount: { type: integer }
    SourceStatus:
      type: object
      description: C7 데이터 출처 표시줄
      properties:
        externalIncluded: { type: boolean }
        lastSyncStatus: { type: string, enum: [ok, failed], nullable: true }
        lastSyncAt: { type: string, format: date-time, nullable: true }
        excludedBuildingCount: { type: integer }
    HistoryEntry:
      type: object
      properties:
        occurredAt: { type: string, format: date-time }
        kind: { type: string, enum: [ai_analysis, field_record, expert_verdict, expert_connection, external] }
        locationText: { type: string, nullable: true }
        defectType: { type: string, nullable: true }
        outcome: { type: string }
        caseId: { type: integer, nullable: true }
        sourceId: { type: integer }
    RepeatPattern:
      type: object
      properties:
        locationText: { type: string }
        defectType: { type: string }
        occurrenceCount: { type: integer }
        basis: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } }
...
    Constant:
      type: object
      properties:
        key: { type: string }
        value: { type: string, nullable: true }
        unit: { type: string, nullable: true }
        description: { type: string }
    AdminUser:
      type: object
      properties:
        userId: { type: integer }
        nameMasked: { type: string }
        emailMasked: { type: string }
        phoneMasked: { type: string, nullable: true }
        orgId: { type: integer, nullable: true }
        orgName: { type: string, nullable: true }
        roles: { type: array, items: { type: string } }
        disabled: { type: boolean }
        mustChangePassword: { type: boolean }
    RoleAssignment:
      type: object
      required: [roles]
      properties:
        roles: { type: array, items: { type: string, enum: [general, facility, building, enterprise, expert, operator] } }
    BuildingAccessGrant:
      type: object
      required: [buildingId, accessKind]
      properties:
        buildingId: { type: integer }
        accessKind: { type: string, enum: [record, manage] }
...